requirements.txt
annotated-types==0.7.0
anyio==4.5.2
attrs==19.3.0
Automat==0.8.0
beautifulsoup4==4.12.3
blinker==1.4
cachetools==5.3.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==23.4.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distlib==0.3.8
distro==1.4.0
distro-info==0.23+ubuntu1.1
dnspython==2.6.1
email-validator==2.2.0
entrypoints==0.3
exceptiongroup==1.2.2
fastapi==0.115.2
fastapi-cli==0.0.5
filelock==3.13.1
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.2
httpx==0.27.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
jinja2==3.1.4
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==4.2.0
netifaces==0.10.4
oauth2client==4.1.3
oauthlib==3.1.0
pexpect==4.6.0
platformdirs==4.2.0
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.18.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.1.2
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
rich==13.9.2
rsa==4.9
SecretStorage==2.3.1
service-identity==18.1.0
shellingham==1.5.4
simplegmail==4.1.1
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
sos==4.5.6
soupsieve==2.5
ssh-import-id==5.10
starlette==0.39.2
systemd-python==234
Twisted==18.9.0
typer==0.12.5
typing-extensions==4.12.2
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
uvicorn==0.31.1
uvloop==0.21.0
virtualenv==20.25.1
wadllib==1.3.3
watchfiles==0.24.0
websockets==13.1
zipp==1.0.0
zope.interface==4.7.1
